sinemapp/requirements.txt
asgiref==3.5.2
Django==3.2.7
django-ckeditor==6.5.1
django-js-asset==2.0.0
pytz==2022.5
sqlparse==0.4.3
